refactor: schema manager is a class instead of interface.
